fix: images now saving properly

+++ LudenWebAPI/Application/Abstractions/Interfaces/Services/IStripeService.cs
@@ -11,8 +11,8 @@ namespace Application.Abstractions.Interfaces.Services
 {
     public interface IStripeService
     {
-        Task<string> CreatePaymentIntentAsync(int userId, int billId);
-        Task<PaymentOrder> CapturePaymentAsync(string paymentIntentId, int userId);
+        Task<string> CreatePaymentIntentAsync(ulong userId, ulong billId);
+        Task<PaymentOrder> CapturePaymentAsync(string paymentIntentId, ulong userId);
         Task<PaymentOrder?> UpdatePaymentStatusAsync(string paymentIntentId, string paymentMethod, string action);
 
     }
